Change GitHub Pages source to 'gh-pages' branch
This change is partly due to the movement to eliminate the term "master"
from computer software, but it is mainly because I wished to avoid
checking out the master branch caused by my muscle memory.

It had been a convention that "master" stands for the main branch of a
Git repository, and the command 'git checkout master' would get you back
to it.

The special branching scheme of this repository adjusted for GitHub
Pages had dictated that 'git checkout master' would not bring you back
to the main branch, which is the 'jekyll' branch, but the branch that
stores the generated static files of this site. These files are only
intended to be used by GitHub Pages and should not be edited directly,
so there was no point to execute 'git checkout master' on this
repository.

I still question whether ditching the term "master" is even helpful in
solving the larger social issue. I believe additional works must be
done, and merely focusing on the nomenclature is an act of avoiding the
real problem. Again, the main goal of this commit is to rename a branch
so its name precisely tells its purpose.