Update gem dependencies and theme files
The Minimal Mistakes theme is now updated to version 4.21.0.  This new
release of the theme contains a few important bug fixes, including one
that resolves improper masthead link collapsing in the navigation bar,
and a patch I authored regarding the figure helper's corrupt HTML DOM
when it is used in a list.