Improve pagination behavior upon untranslated posts
The jekyll-paginate-v2 plugin cannot filter out untranslated posts
automatically and will include them in the post list.  Previously, they
would not be rendered onto the page.  This might lead to generation of
empty pages in the pagination when there is any page where all posts are
not translated.  They would still be in the post list returned by the
paginator; they were merely ignored during the render phase.

The rewritten pagination logic produces the post list on its own by
directly reading from the list of localized posts, and picking the
proper posts according to number of posts per page set in _config.yml
and the page number.  This essentially ensures that the page lists in
pagination will only contain localized posts, so there is no need to
hide any posts when rendering, which eliminates empty pages.