.github/workflows/jekyll.yml: Do not build the site with Jekyll twice
The files being pushed to the gh-pages branch are for the generated
site, so they need not be passed to Jekyll.  Yet GitHub has started to
pass the files pushed to the gh-pages branch to a dynamic workflow that
builds them with Jekyll by default, which is not needed for this
repository and can be disabled to make deployment complete faster.

Reference: https://docs.github.com/en/pages/getting-started-with-github-pages/configuring-a-publishing-source-for-your-github-pages-site#troubleshooting-publishing-problems-with-your-github-pages-site