Only show posts in RSS feed
When a new collection is published, it would clutter the top-level RSS
feed since it usually has multiple child pages.  Those child pages'
entries might look confusing too when they are not presented within the
collection's context in the RSS feed.

If we want to notify the RSS feed subscribers of the new collection, we
can create a redirecting post to the collection's home page, like
posts/2022-08-21-gentoo-config-luks2-grub-systemd.md, so the new
collection appear both in the RSS feed as a single entry and, as an
added bonus, on the website itself's post list.