Configure permalink for section pages
This commit leverages Hugo v0.115.0's new feature to allow a section
landing page inside a parent section to have the same permalink URL
pattern as a normal page, so the headless section workaround is no
longer needed.

The side effect of the configuration is that it changes the permalink of
top-level collection landing pages, but this can be mitigated by
overriding page URL in those pages' front matters.  This workaround was
used when this site still used Jekyll as well.  Since the number of
collection landing pages equals the number of collections on this site
and never exceeds the number of all pages in all collections, this
workaround should not require too much extra work.