Rename `section.html` layout to `collections.html`
`section.html` is the same as `single.html` because it is used only for
this site's collection homepages and subsection parent pages under a
collection, and these pages are to have the same layout as a regular
(i.e. single) page.  However, section pages not under the 'collections'
section are not necessarily supposed to have the same layout.  Thus,
limiting the old `section.html` layout's scope to only the 'collections'
section is better, hence the renaming.