posts/gentoo-config-luks2-grub-systemd: Reformat comment in Markdown
A common suggestion for how to add comments in Markdown is to use the
HTML comment syntax, but Hugo treats it as raw HTML and may emit a
warning.  To eliminate the warning, switch to a hacky way to add
Markdown comments at <https://www.markdownguide.org/hacks/#comments>,
which works and is simple.